refactor(eval): move sweep harness into package
