+++ codeminer/eval/agent_runner/__init__.py
@@ -38,6 +38,16 @@ from .preload import (
     snippet_for_node,
 )
 from .results import summarize_agent_result
+from .sweep import (
+    LANG_GROUP_TO_KEY,
+    all_index_skill_ids,
+    load_dataset_rows,
+    load_full_contexts,
+    run_cell,
+    scenario_for,
+    slug,
+)
+from .sweep_config import SampleConfig, SweepConfig
 from .symbols import (
     build_prebuilt_symbol_span_index,
     build_symbol_span_index,
@@ -49,6 +59,7 @@ from .verify_expand import (
     VerifyExpandRun,
     expansion_seeds_from_candidates,
     graph_verify,
+    load_graph_nav,
     render_expansion,
     run_verify_expand,
 )
@@ -59,11 +70,15 @@ __all__ = [
     "FormatDiagnostics",
     "GATE_SYSTEM",
     "GraphNav",
+    "LANG_GROUP_TO_KEY",
     "PreloadQueryPlan",
+    "SampleConfig",
+    "SweepConfig",
     "Verdict",
     "VerifyExpandRun",
     "VERIFY_SYSTEM_PROMPT",
     "assemble_preload",
+    "all_index_skill_ids",
     "analyze_pareto",
     "build_prebuilt_symbol_span_index",
     "build_symbol_span_index",
@@ -72,11 +87,14 @@ __all__ = [
     "default_agent_skills_dir",
     "expansion_seeds_from_candidates",
     "graph_verify",
+    "load_graph_nav",
     "graph_compose",
     "interleave_ranked",
     "load_agent_skill_contexts",
     "load_cell_jsons",
+    "load_dataset_rows",
     "load_format_diagnostics",
+    "load_full_contexts",
     "load_meaningful_cells",
     "load_pareto_cells",
     "metric_at_k",
@@ -84,11 +102,14 @@ __all__ = [
     "query_is_specific",
     "render_expansion",
     "retrieve_candidates",
+    "run_cell",
     "run_verify_expand",
     "scatter_gather_localize",
     "safe_mean",
     "safe_min",
+    "scenario_for",
     "snippet_for_node",
+    "slug",
     "summarize_agent_result",
     "summarize_format_cells",
     "symbol_leaf",
