+++ scripts/agent_compile/run_sweep.py
@@ -47,17 +47,17 @@ from codeminer.eval.agent_runner.prebuilt import (  # noqa: E402
     repo_path_for,
     stage_prebuilt_indexes,
 )
-from codeminer.eval.agent_runner.symbols import (  # noqa: E402
-    build_prebuilt_symbol_span_index,
-)
-from scripts.agent_compile.lib.config import SweepConfig  # noqa: E402
-from scripts.agent_compile.lib.harness import (  # noqa: E402
+from codeminer.eval.agent_runner.sweep import (  # noqa: E402
     LANG_GROUP_TO_KEY,
     load_dataset_rows,
     load_full_contexts,
     run_cell,
     scenario_for,
     slug,
+)
+from codeminer.eval.agent_runner.sweep_config import SweepConfig  # noqa: E402
+from codeminer.eval.agent_runner.symbols import (  # noqa: E402
+    build_prebuilt_symbol_span_index,
 )
