+++ scripts/retrieval_ablation/__init__.py
@@ -6,6 +6,6 @@
 
 These probes answer "what is the call graph / a given index worth as a
 *retriever*?" — orthogonal to the agent-compile cost study (which is about the
-agent's tool harness). They reuse ``scripts.agent_compile.lib`` for dataset +
+agent's tool harness). They reuse ``codeminer.eval.agent_runner`` for dataset +
 prebuilt-index loading but do not run an agent loop.
 """
