+++ scripts/retrieval_ablation/graph_recall_ablation.py
@@ -95,8 +95,8 @@ def _analyze_instance(
         repo_path_for,
         stage_prebuilt_indexes,
     )
+    from codeminer.eval.agent_runner.sweep import load_dataset_rows, load_full_contexts
     from codeminer.eval.retrieval_eval import collect_targets
-    from scripts.agent_compile.lib.harness import load_dataset_rows, load_full_contexts
 
     rows_by_id, eval_lookup = load_dataset_rows(cfg)
     row = rows_by_id[instance_id]
@@ -130,7 +130,7 @@ def _analyze_instance(
 
 
 def main(argv: Optional[List[str]] = None) -> int:
-    from scripts.agent_compile.lib.config import SweepConfig as SampleConfig
+    from codeminer.eval.agent_runner.sweep_config import SweepConfig as SampleConfig
 
     ap = argparse.ArgumentParser(description=__doc__)
     ap.add_argument("--instances-json", required=True, type=Path)
