+++ scripts/retrieval_ablation/index_compare.py
@@ -66,9 +66,9 @@ def _files(docs: List[Any], idxs: List[int]) -> List[str]:
 
 def _analyze(inst, cfg, prebuilt_dir, cache_root, ks, nlist, nprobe, seeds):
     from codeminer.eval.agent_runner.prebuilt import stage_prebuilt_indexes
+    from codeminer.eval.agent_runner.sweep import load_dataset_rows, load_full_contexts
     from codeminer.eval.retrieval_eval import collect_targets
     from codeminer.graph.roi_subgraph import ROISubgraph
-    from scripts.agent_compile.lib.harness import load_dataset_rows, load_full_contexts
 
     rows_by_id, eval_lookup = load_dataset_rows(cfg)
     row = rows_by_id[inst]
@@ -139,7 +139,7 @@ def _analyze(inst, cfg, prebuilt_dir, cache_root, ks, nlist, nprobe, seeds):
 
 
 def main(argv: Optional[List[str]] = None) -> int:
-    from scripts.agent_compile.lib.config import SweepConfig as SampleConfig
+    from codeminer.eval.agent_runner.sweep_config import SweepConfig as SampleConfig
 
     ap = argparse.ArgumentParser(description=__doc__)
     ap.add_argument("--instances-json", required=True, type=Path)
