+++ test/scripts/test_graphrag_retrieve.py
@@ -68,7 +68,7 @@ def test_retrieve_loads_contexts_from_prebuilt_repo_not_staged_cache(
             )
 
     monkeypatch.setattr(
-        "scripts.agent_compile.lib.harness.load_dataset_rows",
+        "codeminer.eval.agent_runner.sweep.load_dataset_rows",
         fake_load_dataset_rows,
     )
     monkeypatch.setattr(
@@ -76,7 +76,7 @@ def test_retrieve_loads_contexts_from_prebuilt_repo_not_staged_cache(
         fake_stage_prebuilt_indexes,
     )
     monkeypatch.setattr(
-        "scripts.agent_compile.lib.harness.load_full_contexts",
+        "codeminer.eval.agent_runner.sweep.load_full_contexts",
         fake_load_full_contexts,
     )
     monkeypatch.setattr(
